migrate debugger from main editor to sub package
