+++ editor-packages/editor-debugger/index.ts
@@ -1,0 +1,1 @@
+export * from "./debugger-panel";
