+++ editor-packages/editor-debugger/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@code-editor/debugger",
+  "version": "0.0.0",
+  "private": false
+}
