+++ editor/components/index.ts
@@ -2,4 +2,4 @@ export * as figmacomp from "./figma";
 export * as canvas from "./design-preview-as-is";
 export * as code from "./code-editor";
 export * as runner from "./app-runner";
-export * as visualization from "./visualization";
+export * as visualization from "../../editor-packages/editor-debugger/components/visualization";
