+++ editor/next.config.js
@@ -2,6 +2,8 @@ const withTM = require("next-transpile-modules")([
   // region @editor-app
   "@editor-app/live-session",
   "@code-editor/preview-pip", // TODO: remove me. this is for development. for production, use npm ver instead.
+  "@code-editor/debugger",
+  "@code-editor/canvas",
 
   // region editor-submodule deps
   "@base-sdk-fp/auth",
